requirements.txt
pyyaml
jinja2
requests
python-keystoneauth1
python-keystoneclient
python-heatclient
argparse
time
